+++ Sources/Jazz2/LevelHandler.h
@@ -127,7 +127,7 @@ namespace Jazz2
 		void ShowGems(int32_t count) override;
 		StringView GetLevelText(uint32_t textId, int32_t index = -1, uint32_t delimiter = 0) override;
 		void OverrideLevelText(uint32_t textId, const StringView& value) override;
-		void LimitCameraView(float left, float width) override;
+		void LimitCameraView(int left, int width) override;
 		void ShakeCameraView(float duration) override;
 		void SetWeather(WeatherType type, uint8_t intensity) override;
 		bool BeginPlayMusic(const StringView& path, bool setDefault = false, bool forceReload = false) override;
@@ -300,7 +300,7 @@ namespace Jazz2
 		Vector2f _playerRequiredMovement;
 		Vector2f _playerFrozenMovement;
 		bool _playerFrozenEnabled;
-		int32_t _lastPressedNumericKey;
+		uint32_t _lastPressedNumericKey;
 
 		void OnLevelLoaded(const StringView& fullPath, const StringView& name, const StringView& nextLevel, const StringView& secretLevel,
 			std::unique_ptr<Tiles::TileMap>& tileMap, std::unique_ptr<Events::EventMap>& eventMap,
